Use new msp430 crate & use LCD to show temperature
This project was previously not completed and no long compiled in newer
versions of Rust.

This is now brought up to date to Rust edition 2018 and as well relies
on the msp430fr4133 crate which is a peripheral access API generated
with svd2rust & msp430_svd.

MSP430FR4133_TEMPERATURE relies on the uses the msp430_rt crate
for simple system interrupt handling.

This project is fairly low level; There is no HAL for this MCU /
Launchpad board (yet), and little abstractions have been implemented.

Currently functionality:
* Displays the temperature from ADC12 on the LCD.
* Ability to write numbers to the LCD.
* Ability to write strings to the LCD (currently scrolling is not
supported).