Update project dependencies & turn off LTO
Note: As of 2020-07-26 it is necessary to run the latest (2020-06-26)
build of rust nightly to compile this towards msp430-none-elf

Reason being with a change in the linker where the GNU_FRAME_HEADER is
used by default, however msp430's linkers do not have this; this is
fixed in latest nightly.

Also link time optimzations is causing a bug which results in this
project not compiling, so it has been turned off for the time being.